pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zoom-earth-downloader"
version = "1.0.0"
description = "A scheduled service to download satellite tiles from Zoom Earth."
requires-python = ">=3.9"
authors = [
  { name = "lanseria", email = "zhangchao564265135@hotmail.com" },
]
dependencies = [
  "requests>=2.32.3",
  "schedule>=1.2.0",    # 用于定时任务
  "python-dotenv>=1.0.0" # 用于管理环境变量
]

[tool.hatch.build.targets.wheel]
packages = ["src"]